fix: Découpage des cards de activity
