+++ application/social_co2/lib/screens/widgets/CardActivityList.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:social_co2/styles/CardStyles.dart';
+
+class CardActivitiesList extends StatelessWidget {
+  const CardActivitiesList({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.all(20),
+      decoration: primaryCard,
+      child: Row(),
+    );
+  }
+}
