+++ application/social_co2/lib/screens/widgets/CardMoreInformations.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CardMoreInformations extends StatelessWidget {
+  const CardMoreInformations({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return const Text("data2");
+  }
+}
